Move activity indicator to shared
