+++ packages/shared-react/src/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './addresses'
-export * from './addresses/useAddresses'
-export * from './addresses/useUnsortedAddresses'
 export * from './transactions'
+export * from './useActivityIndicator'
